+++ tests/src/unit/express/route.js
@@ -661,6 +661,22 @@ describe('Route', function() {
             };
         });
 
+        it('should throw a Error when invalid target data property argument is provided', function() {
+            var self = this;
+
+            expect(function() {
+                self.route.validate(this.schema, '');
+            }).to.throw(Error);
+        });
+
+        it('should throw a Error when invalid target data property argument is provided (2)', function() {
+            var self = this;
+
+            expect(function() {
+                self.route.validate(this.schema, null);
+            }).to.throw(Error);
+        });
+
         it("should push validation function middleware to the route's stack", function() {
             this.route.validate(this.schema, 'query');